dom/camera/test/test_bug1022766.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for bug 1022766</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="camera_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<video id="viewfinder" width="200" height="200" autoplay></video>
<img src="#" alt="This image is going to load" id="testimage"/>
<script class="testbody" type="text/javascript;version=1.7">

var suite = new CameraTestSuite();

suite.test('bug-1022766', function() {
  function triggerAutoFocus(p) {
    return new Promise(function(resolve, reject) {
      var firstCall = false;
      var secondCall = false;

      function end() {
        if (firstCall && secondCall) {
          resolve();
        }
      }

      // It doesn't matter if the emulator supports focus or not;
      // this is just testing the sequencing.
      suite.camera.autoFocus().then(function(p) {
        ok(false, "First call to autoFocus() succeeded unexpectedly");
        firstCall = true;
        end();
      }, function(e) {
        ok(e.name === 'NS_ERROR_IN_PROGRESS', 'First call to autoFocus() failed with: ' + e);
        firstCall = true;
        end();
      });

      suite.camera.autoFocus().then(function(p) {
        ok(true, "Second call to autoFocus() succeeded");
        secondCall = true;
        end();
      }, function(e) {
        ok(false, "Second call to autoFocus() failed unexpectedly with: " + e);
        secondCall = true;
        end();
      });
    });
  }

  return suite.getCamera()
    .then(triggerAutoFocus, suite.rejectGetCamera)
});

suite.setup()
  .then(suite.run);

</script>
</body>

</html>
